src/requirements.txt
librato-metrics
pyyaml
boto3
